MAINT: move combine_shf_sbf to _three_body.py
